+++ src/OpenTrivia/IOpenTriviaClient.cs
@@ -14,10 +14,20 @@ public interface IOpenTriviaClient
     /// <param name="logger">Optional logger instance for logging diagnostic information. If null, a NullLogger will be used.</param>
     /// <param name="serializer">Optional serializer instance for serializing and deserializing API data. If null, the internal serializer will be used.</param>
     /// <param name="manageRateLimit">Automatically manage rate limits.</param>
+    /// <param name="autoDecode">Automatically decode API responses.</param>
+    /// <param name="encodingType">The encoding type to use for API requests.</param>
+    /// <remarks>
+    /// If an encodingType is provided, it will be used when not explicity provided to method calls.
+    /// If autoDecode is enabled, the client will attempt to decode API responses using the specified encoding type or the default encoding if none is specified. If autoDecode is disabled, the client will return raw API responses without decoding.
+    /// When autoDecode is not enabled (default behavior_, the client will return raw API responses without decoding, regardless of the specified encoding type.
+    /// This allows you to specify the encoding and manage rate limits without automatically decoding responses, giving you more control over how you handle API data.
+    /// </remarks>
     public static IOpenTriviaClient Create(HttpClient httpClient,
         ILogger? logger = null,
         IApiDataSerializer? serializer = null,
-        bool manageRateLimit = false) => new OpenTriviaClient(httpClient, logger, serializer, manageRateLimit);
+        bool manageRateLimit = false,
+        bool autoDecode = false,
+        ApiEncodingType? encodingType = null) => new OpenTriviaClient(httpClient, logger, serializer, manageRateLimit, autoDecode, encodingType);
 
     /// <summary>
     /// Retrieves a new session token from the API.