+++ CHANGELOG.md
@@ -8,13 +8,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Planned
-- Automatic decoding of questions and answers based on encoding type
 - Automatic rate limiting retries
 
 ## [0.9.x]
 ### Added
 - Integration test suite
 - Additional configuration options for `OpenTriviaClient`
+- Automatic decoding of questions and answers based on encoding type
 
 ## [0.9.0] - 2026-02-12